+++ src/assets/level3.json
@@ -1,0 +1,66 @@
+{ "compressionlevel":0,
+ "height":20,
+ "infinite":false,
+ "layers":[
+        {
+         "data":[399, 399, 399, 399, 399, 399, 399, 399, 399, 399, 399, 399, 399, 399, 399, 399, 399, 399, 399, 399, 399, 399, 399, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 399, 399, 534, 399, 399, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 399, 399, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 399, 399, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 399, 399, 399, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 399, 399, 399, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 399, 399, 399, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 399, 399, 399, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 399, 399, 399, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 399, 399, 399, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 399, 399, 399, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 399, 399, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 399, 399, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 399, 399, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 399, 399, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 399, 399, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 399, 399, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 399, 399, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 534, 399, 399, 399, 399, 399, 399, 399, 399, 392, 392, 399, 399, 399, 399, 399, 399, 399, 399, 399, 399, 399],
+         "height":20,
+         "id":3,
+         "name":"bot",
+         "opacity":1,
+         "type":"tilelayer",
+         "visible":true,
+         "width":20,
+         "x":0,
+         "y":0
+        }, 
+        {
+         "data":[0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 392, 392, 0, 0, 0, 0, 0, 0],
+         "height":20,
+         "id":4,
+         "name":"exit",
+         "opacity":1,
+         "type":"tilelayer",
+         "visible":true,
+         "width":20,
+         "x":0,
+         "y":0
+        }, 
+        {
+         "data":[513, 514, 514, 514, 514, 514, 514, 514, 514, 514, 514, 514, 514, 514, 514, 514, 514, 514, 514, 515, 536, 0, 0, 188, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 188, 0, 538, 536, 0, 0, 235, 0, 0, 0, 0, 211, 190, 120, 190, 190, 213, 0, 0, 0, 188, 0, 538, 536, 0, 0, 0, 0, 0, 0, 0, 0, 0, 188, 0, 0, 0, 0, 0, 0, 188, 0, 538, 536, 0, 0, 0, 119, 190, 190, 213, 0, 0, 188, 0, 211, 190, 190, 121, 0, 188, 0, 538, 536, 0, 0, 0, 188, 0, 0, 0, 0, 0, 235, 0, 0, 0, 0, 188, 0, 188, 0, 538, 536, 0, 0, 0, 188, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 188, 0, 235, 0, 538, 536, 0, 189, 0, 188, 0, 0, 211, 190, 190, 190, 190, 190, 213, 0, 188, 0, 0, 0, 538, 536, 0, 188, 0, 188, 0, 0, 0, 553, 555, 0, 553, 555, 0, 0, 188, 0, 0, 0, 538, 536, 0, 165, 190, 188, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 188, 0, 0, 0, 538, 536, 0, 0, 302, 188, 0, 0, 211, 190, 554, 554, 554, 190, 213, 0, 188, 0, 0, 0, 538, 536, 0, 0, 277, 188, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 188, 0, 0, 0, 538, 536, 0, 0, 300, 188, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 188, 0, 189, 0, 538, 536, 0, 0, 0, 165, 190, 190, 213, 0, 119, 120, 121, 0, 211, 190, 167, 0, 188, 0, 538, 536, 0, 0, 0, 0, 0, 0, 0, 0, 142, 145, 144, 0, 0, 0, 0, 0, 188, 0, 538, 536, 0, 189, 0, 0, 0, 0, 0, 0, 142, 145, 144, 0, 0, 0, 0, 0, 188, 0, 538, 536, 0, 188, 0, 0, 0, 189, 0, 0, 142, 145, 144, 0, 0, 189, 0, 0, 235, 0, 538, 536, 0, 188, 0, 0, 0, 188, 0, 0, 142, 145, 144, 0, 0, 188, 0, 0, 0, 0, 538, 536, 0, 188, 0, 0, 0, 188, 0, 0, 142, 145, 144, 0, 0, 188, 0, 0, 0, 0, 538, 559, 560, 166, 560, 560, 560, 167, 0, 0, 142, 145, 144, 0, 0, 165, 560, 560, 560, 560, 561],
+         "height":20,
+         "id":1,
+         "name":"top",
+         "opacity":1,
+         "type":"tilelayer",
+         "visible":true,
+         "width":20,
+         "x":0,
+         "y":0
+        }],
+ "nextlayerid":5,
+ "nextobjectid":1,
+ "orientation":"orthogonal",
+ "renderorder":"right-down",
+ "tiledversion":"1.3.2",
+ "tileheight":32,
+ "tilesets":[
+        {
+         "columns":23,
+         "firstgid":1,
+         "image":"..\/Downloads\/32x32_map_tile v3.1 [MARGINLESS].png",
+         "imageheight":928,
+         "imagewidth":736,
+         "margin":0,
+         "name":"32x32_map_tile v3.1 [MARGINLESS]",
+         "spacing":0,
+         "tilecount":667,
+         "tileheight":32,
+         "tilewidth":32,
+         "transparentcolor":"#ff00ff"
+        }],
+ "tilewidth":32,
+ "type":"map",
+ "version":1.2,
+ "width":20
+}
